requirements.txt
# Core
pydantic>=2.0,<3.0
python-dotenv>=1.0.0

# LLM
google-genai>=1.0.0

# Web Framework (sample app)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Utilities
rich>=13.0.0
